requirements.txt
beautifulsoup4~=4.11.1
python-telegram-bot~=13.12
pytz~=2022.1
requests~=2.28.1
